Rename power_gpio to gpio_configs.
