+++ libopenbmc_intf/Makefile
@@ -5,7 +5,7 @@ PACKAGE_DEPS=gio-unix-2.0 glib-2.0
 SONAME=libopenbmc_intf.so
 VERSION=1
 LIBOBMC=$(SONAME).$(VERSION)
-INCLUDES=openbmc_intf.h openbmc.h gpio.h power_gpio.h
+INCLUDES=openbmc_intf.h openbmc.h gpio.h gpio_configs.h
 
 LDLIBS+=$(shell pkg-config --libs $(PACKAGE_DEPS))
 ALL_CFLAGS+=$(shell pkg-config --cflags $(PACKAGE_DEPS)) -fPIC -Werror $(CFLAGS)
@@ -18,7 +18,7 @@ all: $(SONAME)
 $(SONAME): $(LIBOBMC)
 	ln -sf $^ $@
 
-$(LIBOBMC): lib%.so.$(VERSION): %.o gpio.o power_gpio.o
+$(LIBOBMC): lib%.so.$(VERSION): %.o gpio.o gpio_configs.o
 	$(CC) -shared $(CFLAGS) $(LDFLAGS) -Wl,-soname,$(SONAME) \
 		-o $@ $^ $(LDLIBS)
